Added magic sounds
Magic Controller now has a concept of "layers", which are individual
contributors to the magic status (spells? :). PonyData\GetUsingMagic is
now consistent with an "or" of all these layers. Magic sounds now play
whenever GetUsingMagic returns true. Teleport also received a new sound
to play.

The looped .wav is with many thanks to
https://gist.github.com/jimmcgowan/4268832 ; I have a text file and
project with the sources of the audio and can send them on request.